Fix user

+++ src/main/java/net/study/controller/BoardController.java
@@ -1,8 +1,11 @@
 package net.study.controller;
 
 import net.study.domain.Board;
+import net.study.domain.CurrentUser;
 import net.study.repository.BoardRepository;
 import net.study.util.Paging;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -28,15 +31,14 @@ import java.util.List;
 @RequestMapping(value = "/article")
 public class BoardController {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(BoardController.class);
     @Autowired
     private BoardRepository boardRepository;
 
-    /*
-    게시판 리스트
-     */
     @RequestMapping(value = "/list")
     public String boardList(Model model,
                             @RequestParam(value = "p", required = false) Integer requestPage){
+        LOGGER.debug("Getting board list page for page number={}", requestPage);
 
         if(requestPage == null) requestPage = 1;
 
@@ -52,9 +54,6 @@ public class BoardController {
         Paging paging = new Paging().paging(requestPage, 10, totalCount);
         model.addAttribute("paging", paging);
 
-        /*
-        게시글이 없을 때 수행된다.
-         */
         if(totalCount == 0) {
             model.addAttribute("hasBoard", false);
         } else {
@@ -64,42 +63,34 @@ public class BoardController {
         return "board/list";
     }
 
-    /*
-    게시판 글쓰기 폼
-     */
     @RequestMapping(value = "/write")
     public String boardWrite(Model model) throws Exception{
+        LOGGER.debug("Getting board write form");
 
-//        util.isMemberId(auth);
         return "board/writeForm";
     }
 
-    /*
-    게시판 글쓰기
-     */
     @RequestMapping(value = "/write", method = RequestMethod.POST)
-    public String boardWrite(@RequestParam(value = "title", required = true) String title,
+    public String boardWrite(@ModelAttribute("currentUser")CurrentUser currentUser,
+                             @RequestParam(value = "title", required = true) String title,
                              @RequestParam(value = "content", required = true ) String content){
-
-
-//        String memberId = util.isMemberId(auth);
+        LOGGER.debug("Getting board write for title={}, content={}", title, content);
 
         Board board = new Board();
         board.setTitle(title);
         board.setContent(content);
         board.setPostingDate(new Date());
+        board.setUser(currentUser.getUser());
 
         boardRepository.save(board);
 
         return "redirect:/article/list";
     }
 
-    /*
-    게시판 읽기
-     */
     @RequestMapping(value = "/read/{boardId}")
     public String boardRead(@PathVariable("boardId") Long boardId,
                             Model model){
+        LOGGER.debug("Getting board read page for id={}", boardId);
 
         Board board = boardRepository.findOne(boardId);
 
@@ -110,37 +101,35 @@ public class BoardController {
         return "board/read";
     }
 
-    /*
-    게시판 수정 폼
-     */
     @RequestMapping(value = "/update/{boardId}")
-    public String boardUpdate(@PathVariable("boardId") Long boardId,
+    public String boardUpdate(@ModelAttribute("currentUser")CurrentUser currentUser,
+                              @PathVariable("boardId") Long boardId,
                               Model model){
-
-        //String memberId = util.isMemberId(auth);
+        LOGGER.debug("Getting board update form for id={}", boardId);
 
         Board board = boardRepository.findOne(boardId);
 
-        //util.isEqualMemberId(boardVO.getUserEmail(), memberId);
+        if(!board.checkUser(currentUser.getUser())){
+            return "redirect:/article/read/"+boardId;
+        }
 
         model.addAttribute("board", board);
 
         return "board/updateForm";
     }
 
-    /*
-    게시판 수정
-     */
     @RequestMapping(value = "/update", method = RequestMethod.POST)
-    public String boardUpdate(@RequestParam(value = "boardId", required = true) Long boardId,
+    public String boardUpdate(@ModelAttribute("currentUser")CurrentUser currentUser,
+                              @RequestParam(value = "boardId", required = true) Long boardId,
                               @RequestParam(value = "title", required = true) String title,
                               @RequestParam(value = "content", required = true ) String content){
-
-//        String memberId = util.isMemberId(auth);
+        LOGGER.debug("Getting board update for id={}, title={}, content={}", boardId, title, content);
 
         Board board = boardRepository.findOne(boardId);
 
-//        util.isEqualMemberId(boardVO.getUserEmail(), memberId);
+        if(!board.checkUser(currentUser.getUser())){
+            return "redirect:/article/read/"+boardId;
+        }
 
         if(!board.getTitle().equals(title)) board.setTitle(title);
         if(!board.getContent().equals(content)) board.setContent(content);
@@ -150,18 +139,16 @@ public class BoardController {
         return "redirect:/article/read/"+boardId;
     }
 
-    /*
-    게시판 삭제
-     */
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
-    public String boardDelete(@RequestParam(value = "boardId", required = true) Long boardId){
-
-//        String memberId = util.isMemberId(auth);
-
+    public String boardDelete(@ModelAttribute("currentUser")CurrentUser currentUser,
+                              @RequestParam(value = "boardId", required = true) Long boardId){
+        LOGGER.debug("Getting board delete for id={}", boardId);
 
         Board board = boardRepository.findOne(boardId);
 
-//        util.isEqualMemberId(boardVO.getUserEmail(), memberId);
+        if(!board.checkUser(currentUser.getUser())){
+            return "redirect:/article/read/"+boardId;
+        }
 
         boardRepository.delete(board);
